search: separate result merge logic
